Allow use of designated MouseInjector pointers
Memory addressing beyond the 64MB boundary to support this was
implemented by Ryan Dwyer in the Perfect Dark decomp build system.

This commit leverages this to use designated locations for critical pointers
to values that need to be updated on-the-fly by the MouseInjector.

Tested on gcc, ido

Usage: build the Mouse Injector and Perfect Dark decomp with `MI=1`

Co-authored-by: Catherine Reprobate <ginger.battlequeen@gmail.com>